Revert "Change every id to Long"
This reverts commit c24f39d054b867f089641e6c0939a86d889b2d6e.

+++ api/src/main/java/com/mycelium/local/repository/jsonlog/JsonLog.java
@@ -16,7 +16,7 @@ import io.micronaut.data.model.naming.NamingStrategies;
 public class JsonLog {
     @Id
     @GeneratedValue(GeneratedValue.Type.IDENTITY)
-    public Long id;
+    public Integer id;
     public Date when;
     public boolean operation;
     public String archiveName;
@@ -33,7 +33,7 @@ public class JsonLog {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
